fix(db): re-parent chatgpt identity index migration onto current head
Re-parent 20260814_000000_add_accounts_chatgpt_identity_index onto
20260826_000000_add_http_bridge_event_chunks (current main head) as
20260828_000000_add_accounts_chatgpt_identity_index, keeping a single
linear alembic head. Also recover from an interrupted
CREATE INDEX CONCURRENTLY: detect a leftover indisvalid=false index and
drop it concurrently before rebuilding, matching the repository's other
concurrent-index migrations, so a failed deployment stays retryable
instead of stamping the revision with no usable index.
